IEP-FoodAnalyzer/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
clarifai==9.7.0
pytest-asyncio
pytest
httpx
prometheus_client==0.19.0